move telemetry conversion to sparse file
